feat(article tempelate): add paywall support to article template
